Added a global peers pool which allows for some peers to be used for all torrents #6 (to save overhead of a tracker in certain situations)
This is mainly meant to be used when AT is used as a tightly integrated library
for transferring files among a known group of peers.  For example, I intend to
use it with AdroitGM to distribute the static images needed for the map(s).
In that situation I already know who the clients are, and can tell the peers,
thus saving the overhead of a tracker.